requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
mlflow
pyyaml